Base more plugin-related values on sensible dynamic defaults
Prompting for the "plugin name" FIRST makes sense, since it's the main
piece of information we need to fill in the template.

Later on, we can use that name to suggest sensible default values for
naming the repo, "project", etc. that align with the conventions we use
in the Sopel project/org itself (thus nudging plugin developers toward
those same conventions).